requirements.txt
numpy>=1.21.0
pillow>=8.0.0
keyboard>=0.13.5
scipy>=1.7.0
pywin32>=227
pyautogui>=0.9.54
